updating event and deleting event

+++ src/hooks/useCalendarStore.js
@@ -1,8 +1,9 @@
 import React from 'react'
 import { useDispatch, useSelector } from 'react-redux'
+import Swal from 'sweetalert2'
 import calendarApi from '../api/calendarApi'
 import { convertEventsToDateEvents } from '../helpers'
-import { onAddNewEvent, onDeleteEvent, onSetActiveEvent, onUpdateEvent } from '../store'
+import { onAddNewEvent, onDeleteEvent, onLoadEvents, onSetActiveEvent, onUpdateEvent } from '../store'
 
 export const useCalendarStore = () => {
 
@@ -17,21 +18,41 @@ export const useCalendarStore = () => {
     const startSavingEvent = async(calendarEvent) => {
       // TODO: llegar al backend
 
-      if(calendarEvent._id){
-        // actualizando
-        dispatch(onUpdateEvent({...calendarEvent}))
+      try {
 
-      } else {
-        // creando
-        const {data} = await calendarApi.post('/event', calendarEvent)
-        console.log(data)
-        dispatch(onAddNewEvent({...calendarEvent, id: data.msg.id, user}))
-
+        if(calendarEvent.id){
+          // actualizando
+          await calendarApi.put(`/event/${calendarEvent.id}`, calendarEvent)
+          dispatch(onUpdateEvent({...calendarEvent, user}))
+          return
+  
+        } 
+          // creando
+          const {data} = await calendarApi.post('/event', calendarEvent)
+          console.log(data)
+          dispatch(onAddNewEvent({...calendarEvent, id: data.msg.id, user}))
+      } catch (error) {
+        console.log(error)
+        Swal.fire('Error al guardar', error.response.data.msg, 'error')
       }
+
+     
+
+      
     }
 
     const startDeletingEvent = async() => {
-      dispatch(onDeleteEvent())
+      try {
+        if(activeEvent){
+          await calendarApi.delete(`/event/${activeEvent.id}`)
+          dispatch(onDeleteEvent())
+        }
+      
+      } catch (error) {
+        console.log(error)
+        Swal.fire('Error al eliminar', error.response.data.msg, 'error')
+      }
+     
     }
 
 
@@ -41,6 +62,7 @@ export const useCalendarStore = () => {
         const {data} = await calendarApi.get('/event');
         console.log(data)
         const events = convertEventsToDateEvents(data.eventos)
+        dispatch(onLoadEvents(events))
         console.log(events)
         
       } catch (error) {
@@ -54,7 +76,7 @@ export const useCalendarStore = () => {
     // Properties
     events,
     activeEvent,
-    hasEventSelected: !!activeEvent?._id,
+    hasEventSelected: !!activeEvent?.id,
 
     //Methods
     setActiveEvent,
